backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
openai>=1.40.0
openai-agents>=0.0.1
pydantic-settings>=2.3.0
google-cloud-secret-manager>=2.21.1
